[#2] symlink driver handling

+++ sphinxcontrib/collections/drivers/symlink.py
@@ -20,13 +20,12 @@ This symlink driver can deal with links to folders and files.
 .. code-block:: python
 
     collections = {
-      'my_files: {
+      'my_files': {
          'driver': 'symlink',
          'source': '../extra_files/',
          'target': 'my_data/'
-         }
       }
-   }
+    }
 
 Clean up behavior
 -----------------
@@ -35,38 +34,53 @@ During clean up the symlink gets unlinked/removed.
 """
 
 import os
+import shutil
 
 from sphinxcontrib.collections.drivers import Driver
 
 
 class SymlinkDriver(Driver):
-
     def run(self):
-        self.info('Creating symlink...')
-        source = self.get_path(self.config['source'])
-        target = self.get_path(self.config['target'])
+        self.info("Creating symlink...")
+        source = self.get_path(self.config["source"])
+        target = self.get_path(self.config["target"])
 
         if not os.path.exists(source):
-            self.error('Source {} does not exist'.format(source))
+            self.error("Source {} does not exist".format(source))
             return
 
         try:
-            os.symlink(source, target)
+            # if config['clean'] is not true, symlink exists already
+            if not os.path.exists(os.path.abspath(target)):
+                os.symlink(source, target)
+            else:
+                self.info("Symlink already exists: {}".format(os.path.abspath(target)))
+
+        except FileExistsError as e:
+            # due to paralell builds for different builders (html, pdf, needs,...)
+            # symlinking seems to fail sometimes due to already existing link.
+            self.info(f"Symlink seems to exist already, continue with normal processing.")
+            if not os.path.islink(target):
+                self.error(f"existing target {os.path.abspath(target)} is no symlink")
         except IOError as e:
-            self.error('Problems during creating of symlink.', e)
+            self.error("Problems during creating of symlink.", e)
         except OSError as e:
-            self.error('Problems during creating of symlink. '
-                       'Maybe unprivileged user if running on Windows 10.', e)
+            self.error("Problems during creating of symlink. " "Maybe unprivileged user if running on Windows 10.", e)
 
     def clean(self):
-        source = self.get_path(self.config['source'])
-        target = self.get_path(self.config['target'])
+        source = self.get_path(self.config["source"])
+        target = self.get_path(self.config["target"])
         try:
-            os.unlink(target)
-            self.info('Symlink removed: {}'.format(target))
+            if os.path.exists(os.path.abspath(target)):
+                if not os.path.islink(target):
+                    shutil.rmtree(target)
+                else:
+                    os.unlink(target)
+                self.info("Symlink removed: {}".format(target))
+            else:
+                self.info("Symlink already cleaned: {}.".format(target))
         except FileNotFoundError:
             # Already cleaned? I'm okay with it.
-            self.info('Symlink already cleaned: {}.'.format(target))
+            self.info("Symlink already cleaned: {}.".format(target))
         except IOError as e:
-            self.error('Problems during cleaning for collection {}'.format(self.config['name']), e)
-
+            self.error("Problems during cleaning for collection {}".format(self.config["name"]), e)
